working parking, no waiting time on arrival

+++ Reso2/Readers/Setupcsv.py
@@ -5,8 +5,9 @@ import csv
 # We delete the file first if one already exists
 
 def setupcsv(name="activitylog.csv"):
-    os.remove(name)
+    if(os.path.isdir('/Users/Thomartin/StanleyR/Reso2/%s'%(name))):
+        os.remove(name)
     with open(name,'wt')as log:
         wr = csv.writer(log, quoting=csv.QUOTE_ALL)
-        wr.writerow(["Timestamp","Type","BotID","VehiculeID","ArrivedAt","StartAt","FinisheAt","LeaveAt","TimeBLK","NBdeposit","NbStorage","Nb_Rob_on_Duty","Origin","Destination"])
+        wr.writerow(['Timestamp','Type','BotID','VehiculeID','ArrivedAt','StartAt','FinisheAt','LeaveAt','TimeBLK','NBdeposit','NbStorage','Nb_Rob_on_Duty','Origin','Destination'])
         log.close()
